refactor(runtime): Stage 1 P0 correctness bug fixes
- 1.1: xr_value_from_range: use XR_FROM_PTR to fix uninitialized
  heap_type/flags UB (was reading stack garbage)
- 1.2: XrShape.symbol_to_index: int8_t → int16_t to support field
  index 128+ (int8 overflows to -128, misread as "not found")
- 1.3: xr_array_set_element: replace unsafe XR_TO_FLOAT/XR_TO_INT
  fallback with xr_value_to_int64_coerce/xr_value_to_f64_coerce
  that panic on non-numeric types instead of reading garbage
- 1.4: xr_vm_proto_new: remove 40+ redundant zero-assignments after
  xr_calloc (keep only DYNARRAY_INIT calls)
- 1.5: xr_proto_add_symbol: upgrade warning to XR_CHECK panic when
  symbol count >= 255 (iABC B/C are 8-bit, silent codegen corruption)

+++ src/runtime/object/xshape.c
@@ -101,13 +101,13 @@ static void shape_build_index_table(XrShape *shape) {
     shape->max_symbol = max_sym;
 
     size_t table_size = (size_t)(max_sym - min_sym + 1);
-    shape->symbol_to_index = (int8_t*)xr_malloc(table_size * sizeof(int8_t));
-    memset(shape->symbol_to_index, -1, table_size * sizeof(int8_t));
+    shape->symbol_to_index = (int16_t*)xr_malloc(table_size * sizeof(int16_t));
+    memset(shape->symbol_to_index, -1, table_size * sizeof(int16_t));
 
     for (uint16_t i = 0; i < shape->field_count; i++) {
         SymbolId sym = shape->field_symbols[i];
         XR_DCHECK(sym >= min_sym && sym <= max_sym, "shape_build_index: symbol out of range");
-        shape->symbol_to_index[sym - min_sym] = (int8_t)i;
+        shape->symbol_to_index[sym - min_sym] = (int16_t)i;
     }
 }
